src/mobile_optimization.py
```python
import responsive_design
import touchscreen_input

def optimizeForMobile(dataDisplay, inputField):
    """
    Function to optimize the application for mobile devices.
    It uses the responsive design principles and touchscreen input handling.
    """
    # Apply responsive design to the data display and input field
    responsive_design.applyResponsiveDesign(dataDisplay)
    responsive_design.applyResponsiveDesign(inputField)

    # Handle touchscreen input for the data display and input field
    touchscreen_input.handleTouchscreenInput(dataDisplay)
    touchscreen_input.handleTouchscreenInput(inputField)

    print("Mobile optimization complete.")
```