src/tests/test_mobile_optimization.py
import unittest
from src import mobile_optimization

class TestMobileOptimization(unittest.TestCase):

    def setUp(self):
        self.dataSet = "sample_data"
        self.userInput = "sample_input"
        self.optimizedData = mobile_optimization.optimizeForMobile(self.dataSet, self.userInput)

    def test_optimizeForMobile(self):
        self.assertIsNotNone(self.optimizedData, "Failed to optimize data for mobile devices.")

    def test_dataIntegrity(self):
        self.assertEqual(self.dataSet, self.optimizedData, "Data integrity compromised during mobile optimization.")

    def test_userInputIntegrity(self):
        self.assertEqual(self.userInput, self.optimizedData, "User input integrity compromised during mobile optimization.")

if __name__ == '__main__':
    unittest.main()